Fix amount to support decimals

+++ lib/actions.ts
@@ -10,7 +10,7 @@ export async function createInvoice(formData: FormData) {
   if (!userId) redirect('/login');
 
   const customer = formData.get('customer') as string;
-  const amount = parseInt(formData.get('amount') as string);
+  const amount = parseFloat(formData.get('amount') as string);
   const status = formData.get('status') as string;
 
   await prisma.invoice.create({
